Increase testability and readability of state manager
